bugfix calculating in IndententionFirstLine

+++ DocX/Paragraph.cs
@@ -606,7 +606,7 @@ namespace Novacode
                     XAttribute hanging = ind.Attribute(XName.Get("hanging", DocX.w.NamespaceName));
                     hanging?.Remove();
 
-                    string indentation = ((indentationFirstLine / 0.1) * 57).ToString();
+                    string indentation = ((int)((indentationFirstLine / 0.1) * 57)).ToString();
                     XAttribute firstLine = ind.Attribute(XName.Get("firstLine", DocX.w.NamespaceName));
                     if (firstLine != null)
                         firstLine.Value = indentation;
